Completed the English based FAQ Retrieval Model with all the four factors in their normalized form.
Also, added support to search in Wikipedia, incase the user has a link
to the appropriate wiki page (simply searching the entire wikipedia is
not feasible). For parsing the wikipedia I have written a simple crawler
which should be either spend a day to be improved or be outrightly
replaced with a better crawler.